backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
cryptography>=41.0.0

# Document Processing
langchain>=0.0.300
pypdf>=3.16.0
pdfplumber>=0.10.0
python-docx>=1.0.0
PyPDF2>=3.0.0

# LLM & AI
groq>=0.4.0

# Vector Database - Encrypted Vector Search (Embedded)
cyborgdb-core>=0.1.0

# Embeddings & ML (Local models)
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.21.0
numpy>=1.24.0
# huggingface-hub>=0.19.0  # Keep as backup

# HTTP Clients
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Caching
redis>=5.0.0
hiredis>=2.2.3

# Storage
boto3>=1.28.0
minio>=7.1.17

# Monitoring & Logging
loguru>=0.7.0
prometheus-client>=0.18.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0

# Code Quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Task Queue (Optional)
celery>=5.3.0

